requirements.txt
pyusb>=1.0.2
nose